Add option to BUILD.sh to use external Qt6 libs from Qt Online Installer

+++ BUILD.sh
@@ -90,16 +90,19 @@ fi
     echo "--------------------------------------------------------------------"
     sleep 5
 
+read -p "Build Qt6 from git sources? Select No if using external Qt build: Yes(y) / No(n):- " qt-build
+
+if [ "$qt-build" = "y" ]; then
 ####### Build Qt6 #######
-cd ${SUBMODULES} && git clone https://github.com/qt/qt5.git Qt6
-cd Qt6 && git checkout 6.8.1
-./init-repository --module-subset=qtbase,qtshadertools,qtmultimedia,qtimageformats,qtserialport,qtsvg
-cd .. && mkdir qt6-build && cd qt6-build
-if [ "$choice" = "y" ]; then
-    ${SUBMODULES}/Qt6/configure -prefix /usr/local/js8lib -submodules qtbase,qtshadertools,qtmultimedia,qtimageformats,qtserialport,qtsvg -- -DCMAKE_OSX_ARCHITECTURES="x86_64;arm64"
-else
-    ${SUBMODULES}/Qt6/configure -prefix ${PREFIX} -submodules qtbase,qtshadertools,qtmultimedia,qtimageformats,qtserialport,qtsvg
-fi
+    cd ${SUBMODULES} && git clone https://github.com/qt/qt5.git Qt6
+    cd Qt6 && git checkout 6.8.1
+    ./init-repository --module-subset=qtbase,qtshadertools,qtmultimedia,qtimageformats,qtserialport,qtsvg
+    cd .. && mkdir qt6-build && cd qt6-build
+    if [ "$choice" = "y" ]; then
+        ${SUBMODULES}/Qt6/configure -prefix /usr/local/js8lib -submodules qtbase,qtshadertools,qtmultimedia,qtimageformats,qtserialport,qtsvg -- -DCMAKE_OSX_ARCHITECTURES="x86_64;arm64"
+    else
+        ${SUBMODULES}/Qt6/configure -prefix ${PREFIX} -submodules qtbase,qtshadertools,qtmultimedia,qtimageformats,qtserialport,qtsvg
+    fi
     cmake --build . --parallel
     cmake --install .
     clear
@@ -108,10 +111,15 @@ fi
     echo "--------------------------------------------------------------------"
     sleep 5
 
-cd .. && rm -rf qt6-build
-cd ${SUBMODULES} && git clean -fdx
-git restore *
-cd ..
+    cd .. && rm -rf qt6-build
+    cd ${SUBMODULES} && git clean -fdx
+    git restore *
+    cd ..
+else
+    cd ${SUBMODULES} && git clean -fdx
+    git restore *
+    cd ..
+fi
 
 clear
